Fix sign comparison warning

+++ src/scripting/wrapper.cpp
@@ -1158,7 +1158,7 @@ static SQInteger FloatingImage_get_action_wrapper(HSQUIRRELVM vm)
   try {
     std::string return_value = _this->get_action();
 
-    assert(return_value.size() < std::numeric_limits<SQInteger>::max());
+    assert(return_value.size() < static_cast<size_t>(std::numeric_limits<SQInteger>::max()));
     sq_pushstring(vm, return_value.c_str(), static_cast<SQInteger>(return_value.size()));
     return 1;
 
@@ -1298,7 +1298,7 @@ static SQInteger Gradient_get_direction_wrapper(HSQUIRRELVM vm)
   try {
     std::string return_value = _this->get_direction();
 
-    assert(return_value.size() < std::numeric_limits<SQInteger>::max());
+    assert(return_value.size() < static_cast<size_t>(std::numeric_limits<SQInteger>::max()));
     sq_pushstring(vm, return_value.c_str(), static_cast<SQInteger>(return_value.size()));
     return 1;
 
@@ -2632,7 +2632,7 @@ static SQInteger ScriptedObject_get_action_wrapper(HSQUIRRELVM vm)
   try {
     std::string return_value = _this->get_action();
 
-    assert(return_value.size() < std::numeric_limits<SQInteger>::max());
+    assert(return_value.size() < static_cast<size_t>(std::numeric_limits<SQInteger>::max()));
     sq_pushstring(vm, return_value.c_str(), static_cast<SQInteger>(return_value.size()));
     return 1;
 
@@ -3092,7 +3092,7 @@ static SQInteger ScriptedObject_get_name_wrapper(HSQUIRRELVM vm)
   try {
     std::string return_value = _this->get_name();
 
-    assert(return_value.size() < std::numeric_limits<SQInteger>::max());
+    assert(return_value.size() < static_cast<size_t>(std::numeric_limits<SQInteger>::max()));
     sq_pushstring(vm, return_value.c_str(), static_cast<SQInteger>(return_value.size()));
     return 1;
 
@@ -5602,7 +5602,7 @@ static SQInteger translate_wrapper(HSQUIRRELVM vm)
   try {
     std::string return_value = scripting::translate(arg0);
 
-    assert(return_value.size() < std::numeric_limits<SQInteger>::max());
+    assert(return_value.size() < static_cast<size_t>(std::numeric_limits<SQInteger>::max()));
     sq_pushstring(vm, return_value.c_str(), static_cast<SQInteger>(return_value.size()));
     return 1;
 
@@ -5627,7 +5627,7 @@ static SQInteger __wrapper(HSQUIRRELVM vm)
   try {
     std::string return_value = scripting::_(arg0);
 
-    assert(return_value.size() < std::numeric_limits<SQInteger>::max());
+    assert(return_value.size() < static_cast<size_t>(std::numeric_limits<SQInteger>::max()));
     sq_pushstring(vm, return_value.c_str(), static_cast<SQInteger>(return_value.size()));
     return 1;
 
@@ -5662,7 +5662,7 @@ static SQInteger translate_plural_wrapper(HSQUIRRELVM vm)
   try {
     std::string return_value = scripting::translate_plural(arg0, arg1, static_cast<int> (arg2));
 
-    assert(return_value.size() < std::numeric_limits<SQInteger>::max());
+    assert(return_value.size() < static_cast<size_t>(std::numeric_limits<SQInteger>::max()));
     sq_pushstring(vm, return_value.c_str(), static_cast<SQInteger>(return_value.size()));
     return 1;
 
@@ -5697,7 +5697,7 @@ static SQInteger ___wrapper(HSQUIRRELVM vm)
   try {
     std::string return_value = scripting::__(arg0, arg1, static_cast<int> (arg2));
 
-    assert(return_value.size() < std::numeric_limits<SQInteger>::max());
+    assert(return_value.size() < static_cast<size_t>(std::numeric_limits<SQInteger>::max()));
     sq_pushstring(vm, return_value.c_str(), static_cast<SQInteger>(return_value.size()));
     return 1;
